refactor(core): Split advancedActions.ts into category-based modules for LLM optimization
Reorganize the 817-line advancedActions.ts into a modular directory structure
that enables LLMs to reason about specific card families without loading all
40+ card definitions.

New structure:
- advancedActions/helpers.ts: Effect factory functions (196 lines)
- advancedActions/bolts.ts: Bolt cards - crystal gain/elemental attacks (102 lines)
- advancedActions/red.ts: Red cards - aggressive combat, wound mechanics (202 lines)
- advancedActions/blue.ts: Blue cards - defensive, mana manipulation (186 lines)
- advancedActions/white.ts: White cards - versatile, speed, reputation (185 lines)
- advancedActions/green.ts: Green cards - healing, movement, nature (205 lines)
- advancedActions/dual.ts: Dual-color cards (70 lines)
- advancedActions/index.ts: Re-exports and aggregation (95 lines)

Preserves exact card definitions - only structural reorganization.
All existing exports (ADVANCED_ACTION_CARDS, getAdvancedActionCard) maintained.
