+++ packages/core/src/engine/validActions/cards.ts
@@ -31,7 +31,7 @@ import {
   COMBAT_PHASE_ATTACK,
 } from "../../types/combat.js";
 import { getBasicActionCard } from "../../data/basicActions/index.js";
-import { getAdvancedActionCard } from "../../data/advancedActions.js";
+import { getAdvancedActionCard } from "../../data/advancedActions/index.js";
 import { getSpellCard } from "../../data/spells.js";
 import { DEED_CARD_TYPE_WOUND, DEED_CARD_TYPE_SPELL } from "../../types/cards.js";
 import { describeEffect } from "../effects/describeEffect.js";
